+++ src/canbus/mcp2515_driver.h
@@ -75,7 +75,7 @@ bool mcp2515_driver_send_msg_buffer(uint16_t can_id, uint8_t ext,
 
 bool mcp2515_rx0_is_full(void);
 
-void mcp2515_driver_read_can_message(uint8_t * id, uint8_t * len,
+void mcp2515_driver_read_can_message(uint16_t * id, uint8_t * len,
                                      uint8_t * read_buf);
 
 bool mcp2515_driver_clear_rx0if(void);
